Add Membership join model
The team_id is not nullable, but the user_id is. The reason for this is
that we want the ability to create a "user-like object" that handles
inviting the user, and to also proactively assign resources and/or
manage permissions on optimistically before the actual user object has
been created.