Make membership.user_id not nullable
Initially, the thought was to design this to be nullable based on the
assumption that invitations will hang from the membership class and that
user creation happens at a separate step.

However, devise_invitable hangs the invitation logic on the user. So,
it's better to make sure that user_id can't be nullable.